% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_CCA.R
\name{cca.mai}
\alias{cca.mai}
\title{Leading canonical correlates  based on Mai and Zhang (2019).}
\usage{
cca.mai(xmat, ymat)
}
\arguments{
\item{xmat}{The X matrix, a matrix with n rows and p columns.}

\item{ymat}{The Y matrix, a matrix with n rows and q columns.}
}
\value{
A list. Contains the followings:
\describe{
\item{sa}{Estimator of \eqn{\alpha}, leading left canonical direction.}
\item{sb}{Estimator of \eqn{\beta}, leading right canonical direction.}
\item{srho}{Estimator of leading canonical correlation.}
 }
}
\description{
The code for the general set up
 is provided in Qing Mai's website. We include the special rank one
case here for completeness because we use this SCCA estimator in an example.
}
\details{
The \eqn{l_1} penalty  parameters corresponding to
           \eqn{\alpha} and \eqn{\beta} are
          taken to be \eqn{\sqrt{(\log p)/n}} and
          \eqn{\sqrt{(\log q)/n}}, respectively.
}
\references{
Mai, Q. and Zhang, X. (2019). \emph{An iterative penalized least squares approach to sparse
canonical correlation analysis}, Biometrics, 75, 734-744.
}
